Worked more on error handling. After this I will start with GET related endpoints. I think that after that at least basic programming should be complete

+++ src/main/java/com/JulianGonzalezLopez/Merlin/permissions/PermissionController.java
@@ -5,6 +5,7 @@
 package com.JulianGonzalezLopez.Merlin.permissions;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -24,20 +25,73 @@ public class PermissionController {
     public PermissionController(PermissionServiceInterface permissionService){
         this.permissionService = permissionService;
     }
-    
-    public ResponseEntity<String> create(
+     
+    public ResponseEntity<?> create(
         @RequestBody CreatePermissionRequest createPermissionRequest){
+        
+
+        try{
+            //TYPE ERRORS
+            if(!(createPermissionRequest.getPermission() instanceof Permission)){
+                throw new Error("Permission must be of type permission or capable of transforming into it");
+            }
+
+            //CONTENT RELATED ERRORS
+            if(createPermissionRequest.getUser_id() < 1 && createPermissionRequest.getTable_id() < 1){
+                throw new Error("user_id and table_id must be 1 or higher");
+            }        
+
+            if(createPermissionRequest.getUser_id() < 1){
+                throw new Error("user_id must be 1 or higher");
+            }
+
+            if(createPermissionRequest.getTable_id() < 1){
+                throw new Error("table_id must be 1 or higher");
+            }   
+        }
+        catch(Error e){
+            return new ResponseEntity<>(e.getMessage(), null, HttpStatus.BAD_REQUEST);
+        }
+        
+        //NORMAL WORLFLOW
         permissionService.create(createPermissionRequest);
-        return ResponseEntity
-                .status(201)
-                .build();
+        
+        return new ResponseEntity<>("CREATED", null, HttpStatus.ACCEPTED);
+
     }
             
     public ResponseEntity<String> delete(
         @RequestBody CreatePermissionRequest createPermissionRequest){
-        permissionService.delete(createPermissionRequest);
-        return ResponseEntity
-                .status(200)
-                .build();
+        
+        String error = null;
+
+        //TYPE ERRORS
+        //Checkeo de primitivos realizado por Java
+        if(!(createPermissionRequest.getPermission() instanceof Permission)){
+            error = "Permission must be of type permission or capable of transforming into it";
+        }
+
+        //CONTENT RELATED ERRORS
+        if(createPermissionRequest.getUser_id() < 1 && createPermissionRequest.getTable_id() < 1){
+            error = "user_id and table_id must be 1 or higher";
+        }        
+        
+        if(createPermissionRequest.getUser_id() < 1){
+            error = "user_id must be 1 or higher";
+        }
+        
+        if(createPermissionRequest.getTable_id() < 1){
+            error = "table_id must be 1 or higher";
+        }
+        
+        
+        //ERROR HANDLER
+        if(error != null){
+            return new ResponseEntity<>(error, null, HttpStatus.BAD_REQUEST);
+        }
+              
+        //NORMAL WORLFLOW
+        permissionService.delete(createPermissionRequest);  
+        return new ResponseEntity<>("DELETED", null, HttpStatus.ACCEPTED);
     }
 }
